Crate cricket

Structs

AFL
Provides an interface to one or many underlying afl-fuzz instances

AFLOpts
Contains options/ stateful information about an AFL instance

AFLView
Struct representing an AFL Fuzzer across a network

History
Manages and records a period of data, keeping it stored and sorted

Network
Provides an abstraction over a network of FuzzerViews

Enums

FuzzerError
A set of errors that can occur while dealing with a fuzzer or fuzzerview

Traits

Fuzzer
Trait for types that represent a Fuzzer

FuzzerView
Trait for types that represent a Fuzzer across a network.