Struct cricket::AFL

pub struct AFL {
    // some fields omitted
}

Provides an interface to one or many underlying afl-fuzz instances

Methods

impl AFL

fn new(opts: AFLOpts) -> AFL

Constructs a new AFL struct

Example

let opts = AFLOpts::new(..Default::default());
let afl = AFL::new(opts);

fn get_opts(&self) -> AFLOpts

Returns a copy of the current AFLOpts struct.

Example

// Example of increasing afl instance_count after launch
let mut opts = afl.get_opts();          // Get mutable copy of current opts
opts.instance_count += 2;               // Modify copy accordingly
let mut new_afl = afl::AFL::new(opts);  // Create new AFL instalce with opts
*afl = new_afl;                         //

fn get_opts_ref<'a>(&'a self) -> &'a AFLOpts

Returns a reference to the current AFLOpts struct.

fn get_opts_mut<'a>(&'a mut self) -> &'a mut AFLOpts

Returns a mutable reference to the current AFLOpts struct.

fn get_config(&mut self, argcsv: &str) -> *mut AFL

Unimplemented

fn get_profile(&self, args: &str) -> Vec<Vec<String>>

"$ ./afl-fuzz -i testcase_dir -o sync_dir -S fuzzer03 [...other stuff...]" Based on 'args', generates arguments for AFL Should be moved out of afl.rs to the server Currently only provides a single profile, eventually there will be an 'add_profile