Struct cricket::AFLOpts

pub struct AFLOpts {
    pub afl_path: String,
    pub running: bool,
    pub target_path: String,
    pub whatsup: String,
    pub sync_dir: String,
    pub testcases: String,
    pub scheme: String,
    pub instance_count: usize,
}

Contains options/ stateful information about an AFL instance

Examples

let aflopts = AFLOpts {
    afl_path: "/path/to/afl-fuzz".to_owned(),
    target_path: "/path/to/target".to_owned(),
    ..Default::default()
}

AFLOpts holds the AFL environment data, such as the path to the sync directory, whether AFL is currently running, and the scheme to use when creating fuzzer sync folders.

Fields

afl_path

running

target_path

whatsup

sync_dir

testcases

scheme

instance_count