Struct cricket::History

pub struct History {
    // some fields omitted
}

Manages and records a period of data, keeping it stored and sorted

Methods

impl History

fn new(size: usize) -> History

fn get_average(&self) -> u64

fn push(&mut self, value: Option<u64>)

Takes a value and adds it to the end of the vector. If the vector fills, this history is cleared.

fn get_upper(&self) -> u64

fn get_lower(&self) -> u64

fn save_to_path(&self, path: String)