Struct cricket::Network

pub struct Network {
    // some fields omitted
}

Provides an abstraction over a network of FuzzerViews

Examples

let structure = vec![("workera".into(),vec!["workerb".into()]),("workerb".into(),vec!["workera".into()])];

let mut network = Network::new();

for (host,targets) in structure {
    network.add_worker(
        Box::new(
            AFLView {
                hostname: host,
                neighbors: targets,
                generation: 0,
                pass_rate: 10,
                args: vec!["default".into()]
            })
        )
}

Represents a network of Fuzzers using FuzzerView objects.

Methods

impl Network

fn new() -> Network

Returns a new Network

fn stop(&self)

Commands all workers in the network to end the fuzzing process

fn add_worker(&mut self, view: Box<FuzzerView>)

fn get_worker_scores(&self) -> Vec<Option<u64>>

Writes a serialized json representation of a network to 'path' The network scores itself by adding the score of every Fuzzer it monitors. Returns a vector of optional worker scores. Scores are None when the worker could not be reached

fn collect_scores_interval(&self, lifetime: &u32, interval: &u32, history: &mut History)

Scores all fuzzers within the network based on an interval over the course of a lifetime. Results are stored in history.

Example

// This network will be scored 10 times within a lifetime.
    let lifetime = 1000;
    let interval = 100;
    network_score.collect_scores_interval(&lifetime, &interval, &mut history_handle);

fn fuzz(&self)

Commands remote Fuzzer instances to begin work rust network.fuzz();

fn pass(&self, lifespan: &u32)

Commands FuzzerViews to pass their respective queues to one another based on individual intervals.