Trait cricket::Fuzzer

pub trait Fuzzer {
    fn get_stats(&self) -> Vec<String>;
    fn putq(&self, newq: &BTreeMap<String, String>);
    fn getq(&self) -> Vec<String>;
    fn launch(&mut self, args: &str) -> Result<(), FuzzerError>;
    fn stop(&mut self);
}

Trait for types that represent a Fuzzer

Fuzzers must be able to send and recieve their corpora, provide information on their progress , and be initialized/terminated on command.

Required Methods

fn get_stats(&self) -> Vec<String>

Returns a Vector of Strings representing the work the fuzzer has done.

fn putq(&self, newq: &BTreeMap<String, String>)

Stores newq to disk, newq should represent a map of file names to file content.

fn getq(&self) -> Vec<String>

Returns all of the files in the queue as a BTreeMap of file names to file content.

fn launch(&mut self, args: &str) -> Result<(), FuzzerError>

Begins the fuzzing process. Takes a &str, which can be used as arguments to the fuzzer.

fn stop(&mut self)

Kills all fuzzers